python/pyproject.toml
[project]
name = "verifyvenice"
version = "0.1.0"
description = "Proof-of-concept verification of Venice.ai API claims using zkML"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.12.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "numpy>=1.26",
    "scipy>=1.12",
    "scikit-learn>=1.4",
    "nltk>=3.8",
    "torch>=2.2",
    "onnx>=1.15",
    "onnxruntime>=1.17",
    "pandas>=2.2",
    "tqdm>=4.66",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
